Update Bill of Materials.ods
Filled in details of draft for the Bill of Materials.  At present the Custom PCB may be changed contingent on costs.  A backup plan is to consider implementing on protoboards following the traces for a two layer PCB design described in Eagle CAD free version (only allows top and bottom layers).